Removes explict (float) conversion

+++ World.cpp
@@ -34,7 +34,7 @@ using std::floor;
 // Setup for random real number generator for stars
 std::random_device World::ranDev;
 std::mt19937 World::rng = std::mt19937(ranDev());
-std::uniform_real_distribution<float> World::starDist(0.0,(float)WIDTH);
+std::uniform_real_distribution<float> World::starDist(0.0,WIDTH);
 std::uniform_real_distribution<float> World::enemyStartingVel(-10.,10.);
 std::uniform_real_distribution<float> World::optimalPlayerDist(HEIGHT / 2, HEIGHT - 100);
 std::uniform_int_distribution<int> World::randomInt(-1000, 1000);
@@ -46,7 +46,7 @@ std::uniform_int_distribution<int> World::starBrightness(100, 255);
 void World::makeStar(float startingHeight)
 {
 	// Sets star size, shape, and color
-	Vector2<float> starSize((float)STAR_HEIGHT, (float)STAR_WIDTH);
+	Vector2<float> starSize(STAR_HEIGHT, STAR_WIDTH);
     StarShape newStar(starSize);
 
     // Makes a new star with a random position along width of screen
@@ -76,7 +76,7 @@ void World::updateStars()
     // Move all the stars down
     for(int i = stars.size() - 1; i >= 0; --i)
 	{
-        stars[i].move(0.0, (float)BACKGROUND_SPEED);
+        stars[i].move(0.0, BACKGROUND_SPEED);
         // Move stars to the top with a random width if they reaches the bottom
         if(stars[i].getPosition().y > HEIGHT + STAR_HEIGHT)
 		{
